+++ src/onegov.pay/onegov/pay/__init__.py
@@ -4,7 +4,10 @@ log.addHandler(logging.NullHandler())  # noqa
 
 from onegov.pay.errors import CARD_ERRORS
 from onegov.pay.models import ManualPayment
-from onegov.pay.models import Payable, Payment, PaymentProvider
+from onegov.pay.models import Payable
+from onegov.pay.models import PayableManyTimes
+from onegov.pay.models import Payment
+from onegov.pay.models import PaymentProvider
 from onegov.pay.collections import PaymentCollection, PayableCollection
 from onegov.pay.collections import PaymentProviderCollection
 from onegov.pay.integration import PayApp, process_payment
@@ -16,6 +19,7 @@ __all__ = (
     'CARD_ERRORS',
     'ManualPayment',
     'Payable',
+    'PayableManyTimes',
     'PayableCollection',
     'PayApp',
     'Payment',