+++ src/onegov.pay/onegov/pay/models/__init__.py
@@ -1,5 +1,11 @@
-from onegov.pay.models.payable import Payable
+from onegov.pay.models.payable import Payable, PayableManyTimes
 from onegov.pay.models.payment import ManualPayment, Payment
 from onegov.pay.models.payment_provider import PaymentProvider
 
-__all__ = ('ManualPayment', 'Payable', 'Payment', 'PaymentProvider')
+__all__ = (
+    'ManualPayment',
+    'Payable',
+    'PayableManyTimes',
+    'Payment',
+    'PaymentProvider'
+)